user: Factor out iterators stuff, use for filenames
Iterators in notmuch have a common structure. Factor this out, so that
we don't have to repeat the wrapping code in Clojure. More commonality
could be factored out, but that wouldn't provide much value, I think.
Don't do it for now.

Add a procedure using the new facilities to extract filenames from
messages.